protect posts from other users

+++ app/Http/Controllers/PostsController.php
@@ -7,6 +7,17 @@ use App\Models\Post;
 
 class PostsController extends Controller
 {
+
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -73,6 +84,10 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        
+        if(auth()->user()->id !==$post->user_id){
+            return redirect("/posts")->with('error', 'Unauthorizsed Page');
+        }
         return view("posts.edit")->with('post', $post);
     }
 
@@ -107,6 +122,10 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post =Post::find($id);
+        if(auth()->user()->id !==$post->user_id){
+            return redirect("/posts")->with('error', 'Unauthorizsed Page');
+        }
+        
         $post->delete();
 
         return redirect('/posts')->with('success', 'Post Removed');
